requirements.txt
faiss_cpu==1.9.0.post1
fastapi==0.115.6
numpy==2.2.0
openai==1.58.1
pydantic==2.10.4
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
uvicorn==0.34.0
